src: add whitelist support
This commit adds support for a default whitelist of licenses that are
supported by Red Hat. It also give the ability to add custom license
using a command line option to add to the licenses that are to be
allowed. If a license is not found in one of the above mentioned lists a
warning is printed to the console.