src: support version with dash postfix
Currently, if a dependency has a version with a postfix version number,
for example ups-admin-ui has the version 1.1.4-10 the `-10` was not
taken into account when determining the version.

This commit fixes the above issue and also fixes the issue that this
version was included in both the unknown list (as it does not have a
license that was detected) and in the whitelist. The issue here was that
the while list only inspected the actual license name and ignored the
license file. It now checks the license file and if it is unknown does
not report it as whitelisted, instead the unknown list will report this
license.

Fixes: https://github.com/bucharest-gold/license-reporter/issues/99