src: use license url instead of file path
Instead of using the absolute file url to the license file this commit
uses the url specified in the unifiedlist.json.

This commit also adds a default canonical names mapping file so that one
is not required to be specified on the command line. This is just a
default and can be overridden using the --namemap options as the command
line.

Closes: #110