🎨 Minor suggestions for file names and html output
* 🎨 Put packageName in "Package Artifact" col in html
Before this change, it would be in the "Package Group" column. I guess
the 2 columns would map to a groupId and artifactId in maven. Given
that we don't have groupIds in npm dependencies, I think the name fits
better in the artifact column.

* 🎨 Call it licenses.html instead of license.html

The xml and css (maybe others?) use the plural 'licenses.<extension>'
in their file names. This just changes the html file to match those.